app/requirements.txt
aiogram==3.13.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pydantic-settings==2.5.2
redis[hiredis]==5.1.1
apscheduler==3.10.4
httpx==0.27.2
aiosqlite==0.20.0
